refactor(tui): split tui.rs into a module dir; extract tests
Pure refactor, no behaviour change. Rename src/tui.rs to src/tui/mod.rs and
move the ~2.1k-line test module into src/tui/tests.rs as a child module
(use super::*), which sees the parent's private items unchanged — so no
visibility churn. First step toward issue #73.
